try to avoid infinite diffs (#16)
There are several bugs open against the TF provider
([eg](https://github.com/hashicorp/terraform-provider-aws/issues/9827))
about KFH processor `parameters` not "sticking" and resulting in
infinite diffs. This appears to be something related to not saving the
values in the statefile if they're the defaults.

We're experiencing this internally with `BufferSizeInMBs` trying to be
reset each apply ([internal TF
link](https://app.terraform.io/app/honeycomb/workspaces/prod/runs/run-W9obHXLNwKGNA5gY))

So … let's try not to use the defaults. While we're at it, allow setting
these parameters dynamically, and also sort the keys to try and avoid
any out-of-order plan shenanigans.