add to /examples (#17)
## Which problem is this PR solving?

Fills in /examples with use cases for the top-level module as well as
the individual submodules. I validated each example successfully applies
and gets log data flowing to Honeycomb with the exception of Postgres,
which will require digging into further. (I plan to investigate postgres
this afternoon or first thing monday)

## Short description of the changes

Borrows heavily from our tests to add deployable examples that can be
used as use case documentation and an easy way for users to see how the
integrations work.

## How to verify that this has the expected result

Each example should be successful in deploying to an AWS account and get
logs flowing to Honeycomb.